refactor(step): split step_commands.rs into per-operation modules
Move the 2292-line `src/commands/step_commands.rs` into a `step/` directory
with one file per subcommand:

- `step/commit.rs` — `wt step commit`
- `step/squash.rs` — `wt step squash` (and `handle_squash` reused by `wt merge`)
- `step/rebase.rs` — `wt step rebase` (and `handle_rebase` reused by `wt merge`)
- `step/diff.rs` — `wt step diff`
- `step/copy_ignored.rs` — `wt step copy-ignored`
- `step/promote.rs` — `wt step promote`
- `step/prune.rs` — `wt step prune`
- `step/relocate.rs` — `wt step relocate`
- `step/shared.rs` — `print_dry_run` (commit + squash) and the
  copy-ignored discovery helpers (copy_ignored + promote)
- `step/mod.rs` — re-exports the public API previously exposed by
  `step_commands`

Mechanical extraction with no behavior changes. The bundled file mixed
eight disjoint subcommands sharing almost no code; per-operation modules
make each entry point easier to navigate and modify in isolation. The
public re-export surface in `src/commands/mod.rs` is unchanged.

Also updates two stale path references in comments
(`src/progress.rs`, `tests/integration_tests/step_copy_ignored.rs`) and
extends the stdout-allowlist in `tests/integration_tests/output_system_guard.rs`
to cover the new file paths.
