+++ src/commands/merge.rs
@@ -231,14 +231,14 @@ pub fn handle_merge(opts: MergeOptions<'_>) -> anyhow::Result<()> {
     // --no-hooks vs declined-approval distinction.
     let squashed = if squash_enabled {
         matches!(
-            super::step_commands::handle_squash(
+            super::step::handle_squash(
                 Some(&target_branch),
                 yes,
                 commit_hooks,
                 Some(stage_mode),
                 &mut announcer,
             )?,
-            super::step_commands::SquashResult::Squashed { .. }
+            super::step::SquashResult::Squashed { .. }
         )
     } else {
         false
@@ -248,8 +248,8 @@ pub fn handle_merge(opts: MergeOptions<'_>) -> anyhow::Result<()> {
     let rebased = if rebase {
         // Auto-rebase onto target
         matches!(
-            super::step_commands::handle_rebase(Some(&target_branch))?,
-            super::step_commands::RebaseResult::Rebased { .. }
+            super::step::handle_rebase(Some(&target_branch))?,
+            super::step::RebaseResult::Rebased { .. }
         )
     } else {
         // --no-rebase: verify already rebased, fail if not
